# client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/stops/nearby": {
      "get": {
        "tags": ["stops"],
        "summary": "Near By Stops",
        "operationId": "nearByStops",
        "parameters": [
          {
            "name": "min_lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Min Lat"
            }
          },
          {
            "name": "min_lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Min Lon"
            }
          },
          {
            "name": "max_lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Max Lat"
            }
          },
          {
            "name": "max_lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Max Lon"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Stop"
                  },
                  "title": "Response Nearbystops"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stops/by-name": {
      "get": {
        "tags": ["stops"],
        "summary": "Stops By Name",
        "operationId": "stopsByName",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Stop"
                  },
                  "title": "Response Stopsbyname"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stops/{stop_id}": {
      "get": {
        "tags": ["stops"],
        "summary": "Stop",
        "operationId": "stop",
        "parameters": [
          {
            "name": "stop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Stop Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stop"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stops/{stop_id}/arrival-times": {
      "get": {
        "tags": ["stops"],
        "summary": "Arrival Times",
        "operationId": "arrivalTimes",
        "parameters": [
          {
            "name": "stop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Stop Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArrivalTime"
                  },
                  "title": "Response Arrivaltimes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/routes/{route_id}": {
      "get": {
        "tags": ["routes"],
        "summary": "Route",
        "operationId": "route",
        "parameters": [
          {
            "name": "route_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Route Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Route"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/routes/{route_id}/stops-and-shapes": {
      "get": {
        "tags": ["routes"],
        "summary": "Stops And Shapes",
        "operationId": "stopsAndShapes",
        "parameters": [
          {
            "name": "route_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Route Id"
            }
          },
          {
            "name": "direction_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Direction Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopsAndShapes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/routes/{route_id}/earliest-arrival-times": {
      "get": {
        "tags": ["routes"],
        "summary": "Earliest Arrival Times On Route",
        "operationId": "earliestArrivalTimesOnRoute",
        "parameters": [
          {
            "name": "route_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Route Id"
            }
          },
          {
            "name": "direction_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Direction Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Date"
            }
          },
          {
            "name": "time",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArrivalTimeAtStop"
                  },
                  "title": "Response Earliestarrivaltimesonroute"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/routes/{route_id}/trip-ids": {
      "get": {
        "tags": ["routes"],
        "summary": "Trip Ids For Route",
        "operationId": "tripIdsForRoute",
        "parameters": [
          {
            "name": "route_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Route Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripIds"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{trip_id}": {
      "get": {
        "tags": ["trips"],
        "summary": "Trip",
        "operationId": "trip",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{trip_id}/stop-times": {
      "get": {
        "tags": ["trips"],
        "summary": "Stop Times",
        "operationId": "stopTimes",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StopTime"
                  },
                  "title": "Response Stoptimes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArrivalTime": {
        "properties": {
          "scheduledArrivalTime": {
            "type": "string",
            "title": "Scheduledarrivaltime"
          },
          "updatedArrivalTime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedarrivaltime"
          },
          "trip": {
            "$ref": "#/components/schemas/Trip"
          }
        },
        "type": "object",
        "required": ["scheduledArrivalTime", "trip"],
        "title": "ArrivalTime"
      },
      "ArrivalTimeAtStop": {
        "properties": {
          "stopId": {
            "type": "string",
            "title": "Stopid"
          },
          "stopSequence": {
            "type": "integer",
            "title": "Stopsequence"
          },
          "scheduledArrivalTime": {
            "type": "string",
            "title": "Scheduledarrivaltime"
          },
          "tripId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tripid"
          },
          "updatedArrivalTime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedarrivaltime"
          }
        },
        "type": "object",
        "required": ["stopId", "stopSequence", "scheduledArrivalTime"],
        "title": "ArrivalTimeAtStop"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LineString": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "coordinates": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Coordinates"
          }
        },
        "type": "object",
        "required": ["type", "coordinates"],
        "title": "LineString"
      },
      "Point": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "coordinates": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Coordinates"
          }
        },
        "type": "object",
        "required": ["type", "coordinates"],
        "title": "Point"
      },
...
      "Route": {
        "properties": {
          "routeId": {
            "type": "string",
            "title": "Routeid"
          },
          "agencyId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agencyid"
          },
          "routeShortName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routeshortname"
          },
          "routeLongName": {
            "type": "string",
            "title": "Routelongname"
          },
          "routeColor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routecolor"
          }
        },
        "type": "object",
        "required": ["routeId", "routeLongName"],
        "title": "Route"
      },
...
      "Stop": {
        "properties": {
          "stopId": {
            "type": "string",
            "title": "Stopid"
          },
          "stopCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stopcode"
          },
          "stopName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stopname"
          },
          "stopDesc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stopdesc"
          },
          "stopUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stopurl"
          },
          "wheelchairBoarding": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wheelchairboarding"
          },
          "onStreet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onstreet"
          },
          "atStreet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atstreet"
          },
          "stopLoc": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Point"
              },
              {
                "type": "null"
              }
            ]
          },
          "routes": {
            "items": {
              "$ref": "#/components/schemas/Route"
            },
            "type": "array",
            "title": "Routes",
            "default": []
          }
        },
        "type": "object",
        "required": ["stopId"],
        "title": "Stop"
      },
      "StopTime": {
        "properties": {
          "tripId": {
            "type": "string",
            "title": "Tripid"
          },
          "arrivalTime": {
            "type": "string",
            "title": "Arrivaltime"
          },
          "departureTime": {
            "type": "string",
            "title": "Departuretime"
          },
          "stopId": {
            "type": "string",
            "title": "Stopid"
          },
          "stopSequence": {
            "type": "integer",
            "title": "Stopsequence"
          },
          "pickupType": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pickuptype"
          },
          "dropOffType": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dropofftype"
          },
          "shapeDistTraveled": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shapedisttraveled"
          },
          "timepoint": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timepoint"
          },
          "stop": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Stop"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "tripId",
          "arrivalTime",
          "departureTime",
          "stopId",
          "stopSequence"
        ],
        "title": "StopTime"
      },
...
      "StopsAndShapes": {
        "properties": {
          "stops": {
            "items": {
              "$ref": "#/components/schemas/Stop"
            },
            "type": "array",
            "title": "Stops"
          },
          "shapes": {
            "items": {
              "$ref": "#/components/schemas/LineString"
            },
            "type": "array",
            "title": "Shapes"
          },
          "distinctTrips": {
            "items": {
              "$ref": "#/components/schemas/Trip"
            },
            "type": "array",
            "title": "Distincttrips"
          }
        },
        "type": "object",
        "required": ["stops", "shapes", "distinctTrips"],
        "title": "StopsAndShapes"
      },
      "Trip": {
        "properties": {
          "routeId": {
            "type": "string",
            "title": "Routeid"
          },
          "serviceId": {
            "type": "string",
            "title": "Serviceid"
          },
          "tripId": {
            "type": "string",
            "title": "Tripid"
          },
          "tripHeadsign": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tripheadsign"
          },
          "directionId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Directionid"
          },
          "blockId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blockid"
          },
          "shapeId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shapeid"
          },
          "scheduledTripId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduledtripid"
          },
          "tripShortName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tripshortname"
          },
          "wheelchairAccessible": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wheelchairaccessible"
          },
          "bikesAllowed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bikesallowed"
          },
          "route": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Route"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": ["routeId", "serviceId", "tripId"],
        "title": "Trip"
      },
...
      "TripIds": {
        "properties": {
          "tripIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tripids"
          }
        },
        "type": "object",
        "required": ["tripIds"],
        "title": "TripIds"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}